Add files via upload
Pretty bare bones of my idea. I did this before we had EasyFrame, so I'll likely redo it, hopefully that makes it a bit easier. I'm excited to work on it some more! Even with as little as there is now, I can really see how it will come together!